phase1-foundations/project1-hello-ai/requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
